# api/bindings/api-docs.json
{"openapi":"3.0.3","info":{"title":"snap-shot-api","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/snap-shots":{"get":{"tags":["Snapshot"],"operationId":"handle_get_snapshot_history","responses":{"200":{"description":"Partner account was created","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SnapShotBatch"}}}}}}},"post":{"tags":["Snapshot"],"operationId":"handle_snapshot","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnapShotParams"}}},"required":true},"responses":{"200":{"description":"Partner account was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnapShotBatch"}}}}}}},"/api/snap-shots/{id}":{"get":{"tags":["Snapshot"],"operationId":"handle_get_snapshot_by_id","parameters":[{"name":"id","in":"path","description":"Historical Item Id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Partner account was created","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SnapShotBatch"}}}}}}}},"/jobs":{"get":{"tags":["Batch Job"],"operationId":"handle_get_all_running_jobs","responses":{"200":{"description":"Jobs found successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SnapShotBatchJob"}}}}},"404":{"description":"Jobs not found"}}}},"/jobs/{id}":{"get":{"tags":["Batch Job"],"operationId":"handle_get_job_by_id","parameters":[{"name":"id","in":"path","description":"Job ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Jobs found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnapShotBatchJob"}}}},"404":{"description":"Jobs not found"}}}}},"components":{"schemas":{"CompareImagesReturn":{"type":"object","required":["created_images_paths","deleted_images_paths","diff_images_paths"],"properties":{"created_images_paths":{"type":"array","items":{"type":"string"}},"deleted_images_paths":{"type":"array","items":{"type":"string"}},"diff_images_paths":{"type":"array","items":{"type":"string"}}}},"SnapShotBatch":{"type":"object","required":["id","name","created_at","new_story_book_version","old_story_book_version","diff_images_paths","new_images_paths","old_images_paths"],"properties":{"created_at":{"type":"string","format":"date-time"},"diff_images_paths":{"$ref":"#/components/schemas/CompareImagesReturn"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"new_images_paths":{"type":"array","items":{"type":"string"}},"new_story_book_version":{"type":"string"},"old_images_paths":{"type":"array","items":{"type":"string"}},"old_story_book_version":{"type":"string"}}},"SnapShotBatchJob":{"type":"object","required":["id","status","progress","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"id":{"type":"string","format":"uuid"},"progress":{"type":"number","format":"float"},"snap_shot_batch_id":{"type":"string","format":"uuid","nullable":true},"status":{"$ref":"#/components/schemas/SnapShotBatchJobStatus"},"updated_at":{"type":"string","format":"date-time"}}},"SnapShotBatchJobStatus":{"type":"string","enum":["Pending","Processing","Completed","Failed"]},"SnapShotParams":{"type":"object","properties":{"new":{"type":"string","nullable":true},"old":{"type":"string","nullable":true}}}}},"tags":[{"name":"Batch Job","description":"All about jobs"},{"name":"Snapshot","description":"All about jobs"}]}